refactor: error middleware
